// palcon/web/src/lib/api.ts
export interface GuildMember {
  uid: string;
  name: string;
}

export interface Guild {
  id: string;
  name: string;
  baseCampLevel: number;
  members: GuildMember[];
  memberCount: number;
  bases: { id: string; x: number; y: number }[];
}